Use nix instead of brew for dev container Graphviz
Homebrew was installing a lot of dependencies for it, including
linux-headers, which failed with "fatal: not in a git directory".

Unlike Homebrew, which has bottles, I don't think Nix installs
binary packages, so that is a disadvantage. It might be worthwhile
to get Homebrew working for it and compare the size and time to
install. For now, I'm trying Nix.