Get graphviz from apt in dev container
Nix is heavier and more involved than I'd like, too, and would
require further configuration to make the package available. It is
not obviously better than "manually" compiling Graphviz from source
in the postCreate script, for this use case.

I expect to get a much older version of graphviz when installing
via apt, and this can cause problems:

https://github.com/EliahKagan/codegraph/issues/5

But I'm trying it anyway, because if it works acceptably for
palgoviz, then it will both simplify configuration and speed up
container creation.