Put back graphviz via apt (should work)
It appears I was mistaken in 7044a6a: the powershell feature break
was not due to the apt-packages feature being added. Instead, it
seems to have been a coincidence, where access to that container
was temporarily disrupted.

I was able to produce that with several versions of the dev
container config, including multiple versions that had worked
previously. Now I cannot reproduce it at all, so I am hoping it
was a temporary problem that was fixed or otherwise went away.

This commit brings things back to the way they were in e823fc1.