Have test_greetall.txt request and expect UTF-8
Doctests in test_greetall.txt use subprocess.getstatusoutput to run
the Python interpreter via a shell, testing the behavior of
greetall.py when it is run as a program. This remains the case.

But when I wrote those doctests, I accidentally wrote them in a way
that implicitly assumed the Python subprocess's output encoding
agreed with the encoding that subprocess.getstatusoutput expected.
This is usually, but not always, the case.

The encoding subprocess.getstatusoutput expects is the encoding
expected by the parent process, which calls that to run the child
process. The Python subprocess's output encoding is the encoding
the child process uses for sys.stdout, as used in greetall.py and
any modules it uses, including implicitly via the print builtin.
(It may be relevant that, on Windows, a shell may also affect this,
since when PowerShell is used to process or display text output, it
can change its encoding. Recent versions of PowerShell use UTF-8.)

At some point, this brittle situation was broken by an update, in a
single scenario. The breakage might also have involved an update to
VS Code, outside of which the problem never seems to occur, and/or
the recent update to PowerShell (7.3.2). I'm not sure.

What broke is: Windows, with poetry (not conda/mamba), using pytest
to run the doctests, and using the VS Code test runner (the
"beaker") to run pytest. This seems to be the only scenario that
broke. In particular, I often run all tests on Windows, in a
poetry-managed environment, using pytest. In addition to my doing
that manually, it is one of the ways our tests run via CI on
GitHub. But in VS Code instead of by running the pytest command
manually or from a script, it breaks.

Currently, poetry gives us Python 3.11.1, while conda/mamba gives
us Python 3.11.0. So poetry vs. conda/mamba is probably a proxy for
that distinction.

It *appears* that the cause of the mismatch, broadly speaking, was:

- On Unix-like systems, the parent process expected UTF-8 and
  the child process would output UTF-8. (Good.)

- On Windows, when not run via VS Code, the parent process expected
  Latin-1 and the child process would output Latin-1. (Good.)

- On Windows, with VS Code, under the conditions described above,
  the parent process expected Latin-1 but the child process would
  output UTF-8. (Bad.)

The way the failure manifested was that the "¡" characters in
output from a Spanish greeter (see greet.py and greetall.py) were
received as "Â¡" instead, causing the tests to fail.

Passing encoding='utf-8' to subprocess.getstatusoutput, without
making any other change, is enough to fix the problem in VS Code.
But that breaks the tests when run with the pytest command from a
terminal. This is to say it replaces the bug with a reverse bug.

In any case, I think it is a bug (albeit a less severe one) for
test_greetall.txt to assume the parent and child processes agree on
encoding, even in the absence of a specific situation like this.

This commit fixes the bug by setting the encoding in the
environment the child process inherits (in PYTHONIOENCODING) in
addition to telling subprocess.getstatusoutput to expect UTF-8.