Set shell to fish in yet another place
The chsh command in postCreate, and the VS Code configuration
(terminal.integrated.defaultProfile.linux provided by the fish dev
container feature), were together sufficient for most purposes.

But when using the container *not* in a GitHub Codespace (with the
VS Code dev containers extension), $SHELL is still /bin/bash. Maybe
this will help with that, and address the occasional situations
where bash runs when fish is expected, in non-codespace containers.