Change docs to recommend "pip install -e ."
This changes the top-level readme and the detailed conda/mamba
instructions to recommend running "pip install -e ." instead of
"conda develop .".

This documentation change corresponds to c99cb6f.

See also 0b56e9e. I'm doing this now, rather than as a completely
separate change from the introduction of the dev container config,
because while the dev container config seems unaffected by this bug
when used in a GitHub codespace, it is affected when deployed
locally via the VS Code dev containers extension.

The changes to the poetry instructions are just copyediting. The
actual procedure using poetry is not affected. ("poetry install"
takes care of the corresponding step automatically.)