Prepare to update copyright years
We're using 0BSD, which imposes no restrictions, but it's still
best for stated copyright years to be correct. There are two
changes to make:

(1) Some new files this year have 2023 as their correctly stated
    copyright year, but this year has not been added to LICENSE.

(2) While we're at it, there is actually some material here from
    late 2021, so 2021 should be listed in LICENSE, and in files
    that list copyright years (i.e., those with the full 0BSD
    statement, such as Python modules).

To help find files to update, especially for case (2), I've written
a small script. This commit just adds the script; it does not
update any copyright years. The script is pretty limited and I
think it can be removed in the forthcoming commit that updates the
years (but it may be useful to have it in the commit history).