Update copyright years
As detailed in 1392b94 (the previous commit), we're using 0BSD,
which imposes no restrictions, but it is still best for stated
copyright years to be fully accurate.

This adds 2023 to LICENSE, reflecting substantial additions this
year, which include the configuration and scripts in .devcontainer.

It also corrects an earlier omission of mine where Python modules
we made and committed in late 2021 (rather than 2022) didn't show
that year.

There are also notebooks from December 2021, but we're not showing
full notices with copyright years in notebooks, so those don't have
to be updated.

This commit also removes the script added in 1392b94 that I used to
help me find some of the relevant files.