Temporarily switch to prerelease Python extension
A bug in ms-python.python prevents the extension from loading:

https://github.com/microsoft/vscode/issues/171960

The bug has been fixed but, at this time, the fix is only in the
prerelease version of the extension.

This commit switches to the prerelease version of the extension in
dev containers. I plan to set it back to the stable version, once
it incorporates the bugfix.