Remove forking.md; drop/change references to it
When I made forking.md, it seemed to make sense at the time. I was
describing how to make a codespace for palgoviz, etc.

I don't think my instructions are useful or necessary, compared to
the more detailed GitHub documentation that I was already also
referring to, which, while longer, is more complete and is divided
into sections that makes it pretty accessible.