Carry name and email from host into dev container
This should happen automatically, but it doesn't always, on locally
hosted dev containers.

I'm not sure if this is the best way to do it. If someone used the
dev container on a local folder (instead of cloning in the dev
container), does this modify their local config in the existing
repository's .git directory? If so, this might not be the best way.