Carry in config when repo is *directly* in Windows
This specific situation was not covered before in this repository.

This now covers the hopefully infrequent case where the user
reopens the local directory, in a Windows system, in the container,
rather than cloning the repository in the container.

In that situation, initializeCommand runs *directly* on the host
Windows system. In contrast, in other situations on Windows, it
does not run *directly* on Windows, but instead in the Linux-based
environment that Docker uses do its work. (This is currently either
a WSL 2 system or a Linux-based system running directly in a
Hyper-V virtual machine, depending which backend Docker is
configured to use.)

See:

- https://github.com/EliahKagan/portable-initialize
- https://github.com/EliahKagan/batch/blob/main/initialize.cmd
- https://github.com/EliahKagan/VidDraw/blob/master/.devcontainer/initialize.cmd