Don't carry in core.autocrlf
Currently the repo is set to LF for all text files except batch
files, which applies to all platforms, so it's not necessary to
carry in core.autocrlf.

Since there are arguably some advantages of cloning most files with
native line endings (CRLF on Windows, LF on other systems), and the
original reason for departing from that behavior was the absence of
a technique, like the on here, to carry core.autocrlf in from the
host, that might be undone. If so, then the initialize shell script
will have to be modified correspondingly. Carrying in core.autocrlf
at this time was unintentional and a bug.

Probably the work so far on the ek/container branch should be
merged to main before doing line ending changes affecting how most
files in the repository are handled.

See https://github.com/EliahKagan/palgoviz/pull/45 for context.