Let CodeQL install dependencies, with help
This still uses poetry to export a requirements.txt, but omits the
immediately following step of running pip to install from it. I
believe a CodeQL action will try to do that itself, and that it
will take care of updating the PyPA tools (pip, setuptools, and
wheel) first, which I could do manually but currently am not doing
at all.