Really carry name and email from host into dev container
I meant https://github.com/EliahKagan/palgoviz/pull/55 to do this!
But it turns out the one case where the problem this addresses
usually occurs -- reopening a local folder in a dev container on
Windows -- was a case I had, somehow, failed to test after making
the changes here.

The bug in #55 is that, while the configuration information is
properly gathered by initialize or initialize.cmd, I forgot to add
the code to postCreate to actually *apply* the configuration!

This commit fixes postCreate. It also makes some unrelated small
code-quality improveents in that file.