Support prebuilds
This switches from using a postCreateCommand to using an
onCreateCommand.

Prebuilds (in GitHub Codespaces) could technically already be
used for palgoviz, though this would've conferred no benefit, since
all the slow setup happened in a postCreateCommand which is run in
each new container, rather than being done as part of a prebuild,
and which conferred some disadvantage, since the GitHub Codespaces
interface doesn't show the postCreateCommand log while that command
is running if it is picking up from a prebuild.

An onCreateCommand runs earlier, and we don't need to run any of
the commands in onCreate (renamed from postCreate) that early, but
it should be okay.