"Run some tests" should not run most doctests
Omitting "--doctest-modules" excludes doctests in modules, which is
most doctests in the project.

I'm deliberately still having it run the decorators doctests, and
it will try to run other doctests outside modules if they are added
(which is also intentional). But the greet and greetall doctests
should not run.

This is not an error in the previous commit that renames the
tests; rather, that commit addresses the naming confusion that had
contributed to the CI workflow bug this fixes.