Have CI matrix "os" hold just prefix
And not "-latest", which will appear where "os" is used.

This only applies in the testing workflows. The others are
unchanged, at this time.

The reason is to make job names shorter as shown when viewing a
workflow.

This is only a slight benefit right now, but soon I'll be making
them longer by combining the four testing workflows into two, and
then it will sometimes be important whole job names can usually be
seen when viewing workflows.