Put CI test workflows together into one
This is with minimal refactoring (at least for now).

Only the testing workflows (the ones that run unit tests, using the
pytest, unittest, or doctest test runner utilities) are affected.
Other workflows remain separate (as they should).

The main goal here is to make it easier to look at results of
different kinds of tests together, and to understand the
distinction between the different jobs while examining results.