Update some dependencies (#148)
This updates dependencies but keeps jupyterlab<4 for now.

* Update dependencies

* Put jupyterlab back to 3.6.4

Major version 4 of jupyterlab removes real-time collaboration to a
separate package, jupyter-collaboration. That package currently has
pre-release status on PyPI, and it's not available at all yet on
conda-forge.

Because JupyterLab real-time collaboration is central to the tested
and documented workflow for teaching and learning with palgoviz (to
the extent to which we have such a workflow, I don't think we
should move from stable to prerelease software unnecessarily, at
least not without heavy testing in realistic use cases. That's even
separate from the issue of its unavailability on conda-forge.

It will probably be best to wait until it is both listed as a
regular release (not pre-release status) on PyPI and packaged on
conda-forge.

If the pre-release status persists for a long time but the software
is known to be effectively stable and widely used, and we are able
to test that it behaves well for this application, then it might be
reasonable to move to it even when it is technically a pre-release.

If, for some reason, jupyter-collaboration is unavailable on
conda-forge even after it is considered stable -- though I don't
think that will happen -- then it can of course be listed as a pip
package in environment.yml.

To clarify, jupyterlab 4 itself is considered stable and not
labeled a pre-release. It's the jupyter-collaboration package that
currently has pre-release status on PyPI and is not on conda-forge.