Use JupyterLab 4 (#149)
This upgrades the jupyterlab package to 4.0.2. Since major version
4 of JupyterLab removes the Real-Time Collaboration feature out of
the jupyterlab package and into a new jupyter-collaboration
package, this adds that as well.

That was previously not done before jupyter-collaboration was a
prerelease package at the time major version 4 of jupyterlab was
released. Now jupyter-collaboration is stable, and it is available
in version 1.0.0 from both PyPI and condaforge.

Other indirect dependencies are also upgraded.

This commit supersedes #132.