Update CORS url_end_pattern for port forwarding
https://github.blog/changelog/2023-07-14-codespaces-port-forwarding-domain-name-updates/

This just makes it work both with the current "preview" subdomain,
which will be removed shortly, and without it. By itself, this
should be sufficient to ensure compatibility with the GitHub change
at the end of the month. However, it would be best to eventually
refactor the code to accept an arbitrary domain in a way that both
continues to be clear and avoids characters from the domain (such
as ".") being interpreted as regex metacharacters, and using the
GITHUB_CODESPACES_PORT_FORWARDING_DOMAIN environment variable. This
can be done with "re.escape" in Python, but some refactoring should
probably be done first. This commit doesn't make any such changes.