Update windowed_alt for new sliding_window behavior
As commented, more_itertools.sliding_window no longer supports n=0
as of version 10 of the library. This causes palgoviz.windowed_alt
to fail with ValueError on the first doctest (but not the others).

This changes gencomp1.windowed_alt to treat n=0 specially, while
still otherwisse delegating to moreutils.sliding_window.

Because this no longer *trivially* delegates to sliding_window,
its value as an exercise for researching a library, adding it, and
using it is diminished. At least for now, though, it remains in the
gencomp1 module.