Tweak wording in using-notebooks
Adding raises-exception support may be out of scope for
vscode-jupyter due to lack of widespread interest in the feature.
This slightly rewords the second on VS Code to avoid falsely
implying that the project has committed to adding raises-exception
support or that it will necessarily be added.

Further refinement may be possible in the future, but I'm doing
this small rewording now to avoid potentially confusing inaccuracy.